target

+++ include/main.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: phudyka <phudyka@student.42.fr>            +#+  +:+       +#+        */
+/*   By: dtassel <dtassel@42.nice.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/05 15:49:16 by phudyka           #+#    #+#             */
-/*   Updated: 2023/10/12 10:53:26 by phudyka          ###   ########.fr       */
+/*   Updated: 2023/10/12 07:38:30 by dtassel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,8 +108,8 @@ typedef struct s_params
 
 typedef struct s_sprite
 {
-	double	x;
-	double	y;
+	int	x;
+	int	y;
 	double	distance;
 	double	relative_x;
 	double	relative_y;
